requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
python-dotenv==0.19.2
requests==2.26.0
google-generativeai